+++ src/services/solana/__tests__/nft-metadata-hydrator.spec.js
@@ -93,7 +93,7 @@ describe('hydrate', () => {
 
     await hydrator.hydrate({ ...tritonShaped, uri: 'ipfs://bafyabc' }, {});
 
-    expect(repository.getOffchainMetadata).toHaveBeenCalledWith('https://ipfs.io/ipfs/bafyabc', {});
+    expect(repository.getOffchainMetadata).toHaveBeenCalledWith('https://ipfs.filebase.io/ipfs/bafyabc', {});
   });
 });